Updated Scripture functionality to ES6 classes.
The findScriptures() function is removed and replaced with the ScriptureUtil object that has a parseScriptures() method.  It returns an array of Scripture objects that do inteligent handling of calculating verses.

Scripture object is still not intelligent to the Bible books. It just expects a single word ahead of the chapter:verse chunk of text.